move SimpleHttpServer to a dedicated lib
