+++ libs-scala/http-test-utils/BUILD.bazel
@@ -1,0 +1,20 @@
+# Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_library",
+)
+
+da_scala_library(
+    name = "http-test-utils",
+    srcs = glob(["src/main/scala/**/*.scala"]),
+    scala_deps = [
+    ],
+    tags = ["maven_coordinates=com.daml:http-test-utils:__VERSION__"],
+    visibility = [
+        "//:__subpackages__",
+    ],
+    deps = [
+    ],
+)
