+++ libs-scala/jwt/BUILD.bazel
@@ -63,7 +63,7 @@ da_scala_library(
         ":jwt",
         "//libs-scala/test-evidence/scalatest:test-evidence-scalatest",
         "//libs-scala/test-evidence/tag:test-evidence-tag",
-        "//test-common",
+        "//libs-scala/http-test-utils",
         "@maven//:com_auth0_java_jwt",
         "@maven//:org_scalatest_scalatest_compatible",
     ],
@@ -88,7 +88,7 @@ da_scala_test(
         ":jwt",
         "//libs-scala/test-evidence/scalatest:test-evidence-scalatest",
         "//libs-scala/test-evidence/tag:test-evidence-tag",
-        "//test-common",
+        "//libs-scala/http-test-utils",
         "@maven//:com_auth0_java_jwt",
         "@maven//:org_scalatest_scalatest_compatible",
     ],
