+++ libs-scala/jwt/src/test/scala/com/digitalasset/jwt/JwksSpec.scala
@@ -7,7 +7,6 @@ import com.daml.jwt.domain.{DecodedJwt, Jwt}
 import com.daml.test.evidence.scalatest.ScalaTestSupport.Implicits._
 import com.daml.test.evidence.tag.Security.SecurityTest.Property.Authenticity
 import com.daml.test.evidence.tag.Security.{Attack, SecurityTest}
-import com.daml.testing.SimpleHttpServer
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import scalaz.\/
@@ -15,6 +14,7 @@ import scalaz.syntax.show._
 
 import java.security.KeyPairGenerator
 import java.security.interfaces.{RSAPrivateKey, RSAPublicKey}
+import com.daml.http.test.SimpleHttpServer
 
 class JwksSpec extends AnyFlatSpec with Matchers {
